refactor: test directory, use of size_t for sizes
