+++ CMakeLists.txt
@@ -11,12 +11,8 @@ set(CMAKE_C_FLAGS_RELEASE "${CMAKE_C_FLAGS_RELEASE} -O2")
 
 add_subdirectory(src)
 
-list(APPEND CMAKE_MODULE_PATH "${CMAKE_SOURCE_DIR}/cmake/Modules")
-
 find_package(CMocka CONFIG)
 if (CMocka_FOUND)
-    include(AddCMockaTest)
-    include(AddMockedTest)
     add_subdirectory(tests)
     enable_testing()
 endif(CMocka_FOUND)
