+++ tests/CMakeLists.txt
@@ -1,9 +1,14 @@
-file(GLOB_RECURSE C_SOURCES ${CMAKE_SOURCE_DIR}/src/*.c)
-list(REMOVE_ITEM C_SOURCES ${CMAKE_SOURCE_DIR}/src/main.c)
+file(GLOB_RECURSE SRC_SOURCES ${CMAKE_SOURCE_DIR}/src/*.c)
+list(REMOVE_ITEM SRC_SOURCES ${CMAKE_SOURCE_DIR}/src/main.c)
 
-add_mocked_test(main
-    SOURCES ${C_SOURCES}
-    LINK_LIBRARIES m
-    )
+add_executable(unit_tests unit_tests.c utils_for_tests.c ${SRC_SOURCES})
+add_executable(sequential_tests sequential_tests.c utils_for_tests.c ${SRC_SOURCES})
+add_executable(parallel_tests parallel_tests.c utils_for_tests.c ${SRC_SOURCES})
 
-target_include_directories(test_main PRIVATE ${CMAKE_SOURCE_DIR}/stb_image)
+target_link_libraries(unit_tests PRIVATE m cmocka)
+target_link_libraries(sequential_tests PRIVATE m cmocka)
+target_link_libraries(parallel_tests PRIVATE m cmocka)
+
+add_test(NAME unit_tests COMMAND unit_tests)
+add_test(NAME sequential_tests COMMAND sequential_tests)
+add_test(NAME parallel_tests COMMAND parallel_tests)
