+++ tests/utils_for_tests.c
@@ -1,21 +1,11 @@
-#pragma once
-
 #include "../src/filters/filter.h"
 #include "../src/utils/utils.h"
 
-#define UPPER_SIZE_LIMIT 2500 // To ensure that the filter is not used for too long.
-#define MIN_FACTOR 0.0005
-#define MAX_FACTOR 1.0
+#define STB_IMAGE_IMPLEMENTATION
+#define STB_IMAGE_WRITE_IMPLEMENTATION
 
-/**
- * Initializes an image_rgb structure with the specified dimensions and verifies that
- * memory allocation for its red, green, and blue channels is successful.
- *
- * @param width Image width.
- * @param height Image height.
- *
- * @return An initialized `structure image_rgb`.
- */
+#include "utils_for_tests.h"
+
 struct image_rgb initialize_and_check_image_rgb(int width, int height) {
 	struct image_rgb image = initialize_image_rgb(width, height);
 	assert_non_null(image.red);
@@ -24,15 +14,6 @@ struct image_rgb initialize_and_check_image_rgb(int width, int height) {
 	return image;
 }
 
-/**
- * Creates a random test image with the specified dimensions. Each pixel's red,
- * green, and blue components are assigned random values between 0 and 255.
- *
- * @param width Image width.
- * @param height Image height.
- *
- * @return A `structure image_rgb` representing a randomly generated test image.
- */
 struct image_rgb create_test_image(int width, int height) {
 	struct image_rgb image = initialize_and_check_image_rgb(width, height);
 	for (size_t i = 0; i < (size_t)width * (size_t)height; i++) {
@@ -43,17 +24,6 @@ struct image_rgb create_test_image(int width, int height) {
 	return image;
 }
 
-/**
- * Create a random filter.
- * Core elements >= 0 and bias = 0 to avoid truncating values less than zero and
- * greater than 255.
- *
- * @param size The size of the filter kernel.
- * @param kernel A 2D array representing the kernel matrix. The function populates
- * this array with random values.
- *
- * @return A `struct filter` randomly generated filter.
- */
 struct filter generate_random_filter(int size, double kernel[size][size]) {
 	struct filter random_filter;
 	random_filter.size = size;
@@ -78,16 +48,6 @@ struct filter generate_random_filter(int size, double kernel[size][size]) {
 	return random_filter;
 }
 
-/**
- * Compares two images channel by channel for exact equality.
- *
- * @param expected Pointer to the expected image.
- * @param actual Pointer to the actual image.
- * @param width Image width.
- * @param height Image height.
- *
- * @return true if channels are identical, false otherwise.
- */
 bool compare_channels(struct image_rgb *expected, struct image_rgb *actual,
 					  int width, int height) {
 	for (size_t i = 0; i < (size_t)width * (size_t)height; i++) {
@@ -100,16 +60,6 @@ bool compare_channels(struct image_rgb *expected, struct image_rgb *actual,
 	return true;
 }
 
-/**
- * Compares two images channel by channel with a tolerance of 1.
- *
- * @param expected Pointer to the expected image.
- * @param actual Pointer to the actual image.
- * @param width Image width.
- * @param height Image height.
- *
- * @return true if channels are similar, false otherwise.
- */
 bool compare_channels_with_epsilon(struct image_rgb *expected,
 								   struct image_rgb *actual, int width, int height) {
 	for (size_t i = 0; i < (size_t)width * (size_t)height; i++) {
@@ -122,12 +72,6 @@ bool compare_channels_with_epsilon(struct image_rgb *expected,
 	return true;
 }
 
-/**
- * Print filter details for debugging.
- *
- * @param filter Pointer to the filter.
- * @param name Name of the filter.
- */
 void print_filter(const struct filter *filter, const char *name) {
 	printf("%s:\n", name);
 	printf("{\n");
@@ -151,15 +95,6 @@ void print_filter(const struct filter *filter, const char *name) {
 	printf("\n");
 }
 
-/**
- * Applies zero padding to a convolution filter by embedding the original
- * filter's kernel into the center of a larger kernel filled with zeros.
- *
- * @param padded_filter Pointer to the filter that will receive the zero-padded
- * kernel.
- * @param original_filter Pointer to the original filter whose kernel will be
- * embedded.
- */
 void apply_zero_padding(struct filter *padded_filter,
 						struct filter *original_filter) {
 	// Filling the extended filter kernel with zeros
@@ -179,14 +114,6 @@ void apply_zero_padding(struct filter *padded_filter,
 	}
 }
 
-/**
- * Save an image for debugging failed tests.
- *
- * @param image Image to save.
- * @param width Image width.
- * @param height Image height.
- * @param test_name Name of the test (used for filename).
- */
 void save_image(struct image_rgb image, int width, int height, char *test_name) {
 	unsigned char *result_image = malloc((size_t)width * (size_t)height * (size_t)3);
 	assert_non_null(result_image);
